+++ tests/fixtures/transform_fit_custom.py
@@ -1,8 +1,43 @@
 import pickle
-import pandas as pd
 from scipy.sparse.csr import csr_matrix
 
-from create_transform_pipeline import make_pipeline
+import numpy as np
+import pandas as pd
+from sklearn.compose import ColumnTransformer, make_column_selector
+from sklearn.impute import SimpleImputer
+from sklearn.pipeline import Pipeline
+from sklearn.preprocessing import OneHotEncoder, StandardScaler
+
+
+numeric_selector = make_column_selector(dtype_include=np.number)
+categorical_selector = make_column_selector(dtype_include=np.object)
+
+
+numeric_pipeline = Pipeline(
+    steps=[
+        ("imputer", SimpleImputer(strategy="median", add_indicator=True)),
+        ("scaler", StandardScaler()),
+    ]
+)
+
+categorical_pipeline = Pipeline(
+    steps=[
+        ("imputer", SimpleImputer(strategy="constant", fill_value="missing")),
+        ("onehot", OneHotEncoder(handle_unknown="ignore")),
+    ]
+)
+
+# Sparse preprocessing pipeline, for models such as Ridge that handle sparse input well
+sparse_preprocessing_pipeline = ColumnTransformer(
+    transformers=[
+        ("num", numeric_pipeline, numeric_selector),
+        ("cat", categorical_pipeline, categorical_selector),
+    ]
+)
+
+
+def make_pipeline():
+    return sparse_preprocessing_pipeline
 
 
 def fit(
@@ -57,6 +92,6 @@ def transform(X, transformer, y=None):
     """
     transformed = transformer.transform(X)
     if type(transformed) == csr_matrix:
-        return pd.DataFrame.sparse.from_spmatrix(transformed)
+        return pd.DataFrame.sparse.from_spmatrix(transformed), y
     else:
-        return pd.DataFrame(transformed)
+        return pd.DataFrame(transformed), y