+++ tests/fixtures/transform_fit_custom_no_hook.py
@@ -1,8 +1,42 @@
 import pickle
 import pandas as pd
-from scipy.sparse.csr import csr_matrix
 
-from create_transform_pipeline import make_pipeline
+import numpy as np
+
+from sklearn.compose import ColumnTransformer, make_column_selector
+from sklearn.impute import SimpleImputer
+from sklearn.pipeline import Pipeline
+from sklearn.preprocessing import OneHotEncoder, StandardScaler
+
+numeric_selector = make_column_selector(dtype_include=np.number)
+categorical_selector = make_column_selector(dtype_include=np.object)
+
+
+numeric_pipeline = Pipeline(
+    steps=[
+        ("imputer", SimpleImputer(strategy="median", add_indicator=True)),
+        ("scaler", StandardScaler()),
+    ]
+)
+
+categorical_pipeline = Pipeline(
+    steps=[
+        ("imputer", SimpleImputer(strategy="constant", fill_value="missing")),
+        ("onehot", OneHotEncoder(handle_unknown="ignore")),
+    ]
+)
+
+# Sparse preprocessing pipeline, for models such as Ridge that handle sparse input well
+sparse_preprocessing_pipeline = ColumnTransformer(
+    transformers=[
+        ("num", numeric_pipeline, numeric_selector),
+        ("cat", categorical_pipeline, categorical_selector),
+    ]
+)
+
+
+def make_pipeline():
+    return sparse_preprocessing_pipeline
 
 
 def fit(
@@ -42,21 +76,3 @@ def fit(
     # NOTE: We currently set a 10GB limit to the size of the serialized model or transformer
     with open("{}/artifact.pkl".format(output_dir), "wb") as fp:
         pickle.dump(transformer, fp)
-
-
-def transform(X, transformer, y=None):
-    """
-    Parameters
-    ----------
-    X: pd.DataFrame - training data to perform transform on
-    transformer: object - trained transformer object
-    y: pd.Series (optional) - target data to perform transform on
-    Returns
-    -------
-    transformed DataFrame resulting from applying transform to incoming data
-    """
-    transformed = transformer.transform(X)
-    if type(transformed) == csr_matrix:
-        return pd.DataFrame.sparse.from_spmatrix(transformed)
-    else:
-        return pd.DataFrame(transformed)