+++ tests/fixtures/transform_fit_custom_sparse_input.py
@@ -1,8 +1,6 @@
 import pickle
 import pandas as pd
-from scipy.sparse.csr import csr_matrix
-
-from create_transform_pipeline import make_pipeline
+from sklearn.decomposition import TruncatedSVD
 
 
 def fit(
@@ -13,12 +11,9 @@ def fit(
 ):
     """
     This hook must be implemented with your fitting code, for running drum in the fit mode.
-
     This hook MUST ALWAYS be implemented for custom training models. For custom transformers, the
     transform hook below is also required.
-
     For inference models, this hook can stick around unimplemented, and won’t be triggered.
-
     Parameters
     ----------
     X: pd.DataFrame - training data to perform fit on
@@ -26,13 +21,12 @@ def fit(
     output_dir: the path to write output. This is the path provided in '--output' parameter of the
         'drum fit' command.
     kwargs: Added for forwards compatibility
-
     Returns
     -------
     Nothing
     """
-    transformer = make_pipeline()
-    transformer.fit(X, y)
+    transformer = TruncatedSVD()
+    transformer.fit(X)
 
     # You must serialize out your transformer to the output_dir given, however if you wish to change this
     # code, you will probably have to add a load_model method to read the serialized model back in
@@ -42,21 +36,3 @@ def fit(
     # NOTE: We currently set a 10GB limit to the size of the serialized model or transformer
     with open("{}/artifact.pkl".format(output_dir), "wb") as fp:
         pickle.dump(transformer, fp)
-
-
-def transform(X, transformer, y=None):
-    """
-    Parameters
-    ----------
-    X: pd.DataFrame - training data to perform transform on
-    transformer: object - trained transformer object
-    y: pd.Series (optional) - target data to perform transform on
-    Returns
-    -------
-    transformed DataFrame resulting from applying transform to incoming data
-    """
-    transformed = transformer.transform(X)
-    if type(transformed) == csr_matrix:
-        return pd.DataFrame.sparse.from_spmatrix(transformed)
-    else:
-        return pd.DataFrame(transformed)